Updated flight data, with some historical details included
In order to try to address seasonality, flight data now includes historical
services, limited to 12 months although currently there is only data since
April 2011.   Of course the downside of this is that services which are
permanently discontinued (i.e. not just seasonally) will stick around for
up to a year.   Anything major can still be adjusted by adding a definition
in the overrides.txt file